reorganize DataGrid components
